+++ grails-app/controllers/io/xh/hoist/admin/ConfigDiffAdminController.groovy
@@ -10,7 +10,7 @@ package io.xh.hoist.admin
 import grails.gorm.transactions.ReadOnly
 import io.xh.hoist.BaseController
 import io.xh.hoist.config.AppConfig
-import io.xh.hoist.json.JSON
+import io.xh.hoist.json.JSONParser
 import io.xh.hoist.security.Access
 
 @Access(['HOIST_ADMIN'])
@@ -26,7 +26,7 @@ class ConfigDiffAdminController extends BaseController {
 
     def applyRemoteValues() {
         def records = params.get('records')
-        configDiffService.applyRemoteValues(JSON.parse(records))
+        configDiffService.applyRemoteValues(JSONParser.parseArray(records))
 
         renderJSON(success: true)
     }