+++ grails-app/services/io/xh/hoist/monitor/MonitoringService.groovy
@@ -13,7 +13,6 @@ import grails.gorm.transactions.ReadOnly
 import io.xh.hoist.BaseService
 import io.xh.hoist.util.Timer
 import io.xh.hoist.async.AsyncSupport
-import org.grails.web.json.JSONObject
 
 import java.util.concurrent.ConcurrentHashMap
 
@@ -212,8 +211,8 @@ class MonitoringService extends BaseService implements AsyncSupport, EventPublis
         return monitorConfig.monitorTimeoutSecs ?: 15
     }
 
-    private JSONObject getMonitorConfig() {
-        configService.getJSONObject('xhMonitorConfig')
+    private Map getMonitorConfig() {
+        configService.getMap('xhMonitorConfig')
     }
 
     void clearCaches() {