+++ src/main/groovy/io/xh/hoist/http/JSONClient.groovy
@@ -9,26 +9,24 @@ package io.xh.hoist.http
 
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
-import io.xh.hoist.json.JSON
+import io.xh.hoist.exception.ExternalHttpException
+import io.xh.hoist.json.JSONParser
 import org.apache.http.client.methods.CloseableHttpResponse
 import org.apache.http.client.methods.HttpRequestBase
 import org.apache.http.impl.client.CloseableHttpClient
 import org.apache.http.impl.client.HttpClients
 import org.apache.http.protocol.BasicHttpContext
-import org.grails.web.json.JSONArray
-import org.grails.web.json.JSONObject
 
 import static org.apache.http.HttpStatus.SC_NO_CONTENT
 import static org.apache.http.HttpStatus.SC_OK
 
 /**
- * A thin wrapper around an Apache HttpClient to execute HTTP requests and return pre-parsed
- * JSONObjects and JSONArrays. (Can also return responses as text or an int HTTP status code.)
+ * A thin wrapper around an Apache HttpClient to execute HTTP requests and return parsed
+ * java objects representing the JSON content.  This object can also return responses as
+ * text or an int HTTP status code.
  *
  * This wrapper also provides additional customizations around exceptions. If a response returns
  * with an HTTP status code indicating an error (i.e. *not* 200-204), the execute method will throw.
- * This class attempt to decode JSON serialized Hoist or java.lang exceptions from a remote host
- * and rethrow the actual exception class if possible, otherwise falling back to a RuntimeException.
  */
 @Slf4j
 @CompileStatic
@@ -46,67 +44,100 @@ class JSONClient {
     //--------------------------------------------------
     // Convenience Methods for getting decoded response
     //---------------------------------------------------
-    JSONObject executeAsJSONObject(HttpRequestBase method) {
-        String ret = executeAsString(method)
-        return ret ? (JSONObject) JSON.parse(ret) : null
-    }
-
-    JSONArray executeAsJSONArray(HttpRequestBase method) {
-        String ret = executeAsString(method)
-        return ret ? (JSONArray) JSON.parse(ret) : null
-    }
-
-    String executeAsString(HttpRequestBase method) {
-        def response
+    /**
+     * Execute and parse a request expected to return a single JSON object.
+     */
+    Map executeAsMap(HttpRequestBase method) {
+        def response = null
         try {
             response = execute(method)
             def statusCode = response.statusLine.statusCode
             if (statusCode == SC_NO_CONTENT) return null
-            return response.entity.content.getText('UTF-8')
+            return JSONParser.parseObject(response.entity.content)
         } finally {
-            if (response) response.close()
+            response?.close()
         }
     }
 
+    /**
+     * Execute and parse a request expected to return an array of JSON objects.
+     */
+    List executeAsList(HttpRequestBase method) {
+        def response = null
+        try {
+            response = execute(method)
+            def statusCode = response.statusLine.statusCode
+            if (statusCode == SC_NO_CONTENT) return null
+            return JSONParser.parseArray(response.entity.content)
+        } finally {
+            response?.close()
+        }
+    }
+
+    /**
+     * Execute request and return raw content string.
+     */
+    String executeAsString(HttpRequestBase method) {
+        def response = null
+        try {
+            response = execute(method)
+            def statusCode = response.statusLine.statusCode
+            if (statusCode == SC_NO_CONTENT) return null
+            return response.entity.content.text
+        } finally {
+            response?.close()
+        }
+    }
+
+    /**
+     * Execute request and return status code only.
+     */
     Integer executeAsStatusCode(HttpRequestBase method) {
-        def response
+        def response = null
         try {
             response = execute(method)
             return response.statusLine.statusCode
         } finally {
-            if (response) response.close()
+            response?.close()
         }
     }
 
-    
+
     //------------------------
     // Implementation
     //------------------------
     private CloseableHttpResponse execute(HttpRequestBase method) {
         CloseableHttpResponse ret = null
-        Exception cause = null
-        String statusText
+        Throwable cause = null
+        Integer statusCode = null
+        String statusText = ''
         boolean success
+
+        if (!method.containsHeader('Content-type') && method.method in ['POST', 'PUT', 'PATCH']) {
+            method.setHeader('Content-type', 'application/json')
+        }
+
+        if (!method.containsHeader('Accept')) {
+            method.setHeader('Accept', 'application/json')
+        }
 
         try {
             ret = executeRaw(_client, method)
-            Integer statusCode = ret.statusLine.statusCode
+            statusCode = ret.statusLine.statusCode
             success = (statusCode >= SC_OK  && statusCode <= SC_NO_CONTENT)
-            statusText = statusCode.toString()
-        } catch (Exception e) {
+            if (!success) {
+                cause = parseException(ret)
+                statusText = statusCode.toString()
+            }
+        } catch (Throwable e) {
             cause = e
             statusText = e.message
             success = false
         }
 
         if (!success) {
-            def exception = ret ? parseException(ret) : null
-            if (!exception) {
-                def msg = "Failure calling ${method.URI} : $statusText"
-                exception = new RuntimeException(msg, cause)
-            }
-            if (ret) ret.close()
-            throw exception
+            ret?.close()
+            throw new ExternalHttpException("Failure calling ${method.URI} : $statusText", cause, statusCode)
         }
 
         return ret
@@ -118,12 +149,10 @@ class JSONClient {
 
     // Attempt to parse a reasonable exception from failed response, but never actually throw if not possible.
     private Throwable parseException(CloseableHttpResponse response){
-        def text = response.entity.content.getText()
-        if (!text) return null
         try {
-            def ex = (JSONObject) JSON.parse(text),
-                className = ex.className,
-                msg = ex.message
+            def ex = JSONParser.parseObject(response.entity.content),
+                className = ex?.className,
+                msg = ex?.message
 
             // Message is required
             if (msg instanceof String) {